Fixed the Repo
Had some issues. Also, added leveling of the graph - the low-end is
reduced visually some, and the high-end is increased, so the leveling
should be about even throughout. It's not a logarithmic curve, as it
probably should be, but who cares? Oh, and I added a couple music files.
Enjoy.